refactor docs && structs
