--- src/observers/mod.rs
@@ -1,7 +1,0 @@
-mod alive;
-mod log;
-mod observer;
-
-pub use alive::AliveTracker;
-pub use log::LogWriter;
-pub use observer::Observer;
